renamed txgenerator
