+++ src/util/static/intPatch.ts
