+++ tslint.json
@@ -1,7 +1,8 @@
 {
-    "extends": "tslint:recommended",
-    "rules": {
-        "no-var-requires": false,
-        "trailing-comma": [false]
+    "extends":"tslint:recommended",
+    "rules":{
+        "no-var-requires":false,
+        "trailing-comma":[false],
+        "prefer-const":false
     }
 }
